InputMismatchException error solved

+++ src/BasicApporach/CountDigits.java
@@ -6,11 +6,11 @@ public class CountDigits {
     public static void main(String[] args){
         Scanner sc = new Scanner(System.in);
         System.out.print("Enter a number: ");
-        int num = sc.nextInt();
+        long num = sc.nextLong();
         System.out.println("Digit count in "+num+ " is: "+countNumberDigit(num));
     }
 
-    public static int countNumberDigit(int num){
+    public static int countNumberDigit(long num){
         int res = 0;
         if(num == 0)
             return 1;
